Refactor data collectors to take file directory
This allows for a very easy file system management. We specify the actual data file in a constant variable and prepend the given data directory to this.

If we need to change where we store the data, it is simple to change the data root directory and keep consistent file naming.